The work can be split between a number of workers

+++ experiment.py
@@ -2,20 +2,26 @@ from collections import OrderedDict
 
 from project.augmentation import FakeAugmenter, CombinerAugmenter
 from project.data import FileDataset
-from project.helpers import get_options_combinations, log_scale
+from project.helpers import get_options_combinations, log_scale, get_index_and_increment
 from project.network.lstm import LSTMNetwork
 from project.vectorization.embedding import WordEmbedding
 from datetime import datetime
 
-import numpy as np
 import csv
+import os
+import random
 
+base_path = './'
 
-results_file = 'results-trec.csv'
+results_file = base_path + 'results.csv'
+worker_index_file = base_path + 'workers.index'
 
-reduction_granularity = 30
-repeat_times = 20
+reduction_granularity = 30      # How many values between 0 and 1 for `r`
+repeat_times = 20               # How many times will each experiment be repeated
+one_worker_per_time = True      # Should I use one worker per each time?
 
+worker_id = random.randint(1000, 9999)
+print("worker_id=%d" % worker_id)
 
 
 datasets = {
@@ -39,13 +45,36 @@ options.update({"reduction": log_scale(steps=reduction_granularity),})
 options.update({"time": list(range(repeat_times)),})
 
 combinations = get_options_combinations(options)
+print("Generated a list of %d combinations for the experiment." % len(combinations))
 
-header = ["Start time", "End time", "Dataset", "Augmentation technique", "Dataset Reduction (r)",
+if one_worker_per_time:
+    print("Multi-worker experiment. Using a section of the list.")
+    this_index = get_index_and_increment(worker_index_file)
+    if this_index >= repeat_times:
+        print("WARNING. Obtained index %d, but the experiment only needs %d repetitions. Terminating" \
+              % (this_index, repeat_times))
+        exit(0)
+    print("worker_index=%d" % this_index)
+    list_no = int(len(combinations) / repeat_times)
+    list_start = int(this_index * list_no)
+
+else:
+    print("Single-worker experiment. Using entire list of combinations.")
+    list_start = 0
+    list_no = len(combinations)
+
+print("Slicing the combinations. start_i=%d, no_elements=%d." % (list_start, list_no))
+
+combinations = combinations[list_start:(list_start + list_no)]
+
+header = ["Worker ID",
+          "Start time", "End time", "Dataset", "Augmentation technique", "Dataset Reduction (r)",
           "Experiment #", "Training Epochs", "Test loss", "Test accuracy"]
 
-with open(results_file, 'w') as csvfile:
-    writer = csv.writer(csvfile)
-    writer.writerow(header)
+if not os.path.exists(results_file):
+    with open(results_file, 'w') as csvfile:
+        writer = csv.writer(csvfile)
+        writer.writerow(header)
 
 for i, options in enumerate(combinations):
 
@@ -77,7 +106,8 @@ for i, options in enumerate(combinations):
 
     end_time = datetime.now()
 
-    line = [start_time, end_time,
+    line = [worker_id,
+            start_time, end_time,
             options['dataset'],
             str(options['augmenter']),
             options['reduction'],
